renamed mongodb to mongodblocalclient to stop it being confused
