+++ scripts/initialize_db.py
@@ -6,7 +6,7 @@
 import os
 import shutil
 
-from mongodb import MongoDbLocalClient
+from mongodblocalclient import MongoDbLocalClient
 
 if __name__ == "__main__":
     print("################################################################################")
